requirements.txt
hikari>=2.0.0.dev82
hikari-lightbulb>=0.0.49